Fix code duplications
