+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/performance/AvoidDmlStatementsInLoopsRule.java
@@ -24,6 +24,7 @@ public class AvoidDmlStatementsInLoopsRule extends AbstractAvoidNodeInLoopsRule 
         setProperty(CODECLIMATE_BLOCK_HIGHLIGHTING, false);
     }
 
+    // CPD-OFF - the same visits are in the replacement rule OperationWithLimitsInLoopRule
     @Override
     public Object visit(ASTDmlDeleteStatement node, Object data) {
         return checkForViolation(node, data);
@@ -53,4 +54,5 @@ public class AvoidDmlStatementsInLoopsRule extends AbstractAvoidNodeInLoopsRule 
     public Object visit(ASTDmlUpsertStatement node, Object data) {
         return checkForViolation(node, data);
     }
+    // CPD-ON
 }
