+++ pmd-javascript/src/main/java/net/sourceforge/pmd/lang/ecmascript/ast/ASTNewExpression.java
@@ -8,7 +8,7 @@ import org.mozilla.javascript.ast.NewExpression;
 
 import net.sourceforge.pmd.annotation.InternalApi;
 
-public class ASTNewExpression extends AbstractEcmascriptNode<NewExpression> {
+public class ASTNewExpression extends AbstractFunctionCallNode<NewExpression> {
     @Deprecated
     @InternalApi
     public ASTNewExpression(NewExpression newExpression) {
@@ -18,22 +18,6 @@ public class ASTNewExpression extends AbstractEcmascriptNode<NewExpression> {
     @Override
     public Object jjtAccept(EcmascriptParserVisitor visitor, Object data) {
         return visitor.visit(this, data);
-    }
-
-    public EcmascriptNode<?> getTarget() {
-        return (EcmascriptNode<?>) getChild(0);
-    }
-
-    public int getNumArguments() {
-        return node.getArguments().size();
-    }
-
-    public EcmascriptNode<?> getArgument(int index) {
-        return (EcmascriptNode<?>) getChild(index + 1);
-    }
-
-    public boolean hasArguments() {
-        return getNumArguments() != 0;
     }
 
     public boolean hasInitializer() {
